requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
Pillow==8.2.0
psycopg2==2.9.1
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1